+++ src/black_holes/EAGLE/black_holes_struct.h
@@ -27,7 +27,7 @@
 struct black_holes_part_data {
 
   /*! ID of the black-hole that will swallow this #part. */
-  long long swallow_id;
+  int64_t swallow_id;
 };
 
 /**
@@ -36,7 +36,7 @@ struct black_holes_part_data {
 struct black_holes_bpart_data {
 
   /*! ID of the black-hole that will swallow this #bpart. */
-  long long swallow_id;
+  int64_t swallow_id;
 
   /*! Mass of the black-hole that will swallow this #bpart. */
   float swallow_mass;
@@ -71,7 +71,7 @@ black_holes_mark_part_as_swallowed(struct black_holes_part_data* p_data) {
  *
  * @param p_data The #part's #black_holes_part_data structure.
  */
-__attribute__((always_inline)) INLINE static long long
+__attribute__((always_inline)) INLINE static int64_t
 black_holes_get_part_swallow_id(struct black_holes_part_data* p_data) {
 
   return p_data->swallow_id;
@@ -108,7 +108,7 @@ black_holes_mark_bpart_as_merged(struct black_holes_bpart_data* p_data) {
  *
  * @param p_data The #bpart's #black_holes_bpart_data structure.
  */
-__attribute__((always_inline)) INLINE static long long
+__attribute__((always_inline)) INLINE static int64_t
 black_holes_get_bpart_swallow_id(struct black_holes_bpart_data* p_data) {
 
   return p_data->swallow_id;
