+++ src/stars/GEAR/stars_part.h
@@ -35,7 +35,7 @@
 struct spart {
 
   /*! Particle ID. */
-  long long id;
+  int64_t id;
 
   /*! Pointer to corresponding gravity part. */
   struct gpart* gpart;
@@ -116,13 +116,13 @@ struct spart {
   int num_ngb_density;
 
   /*! List of interacting particles in the density SELF and PAIR */
-  long long ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS_STARS];
+  int64_t ids_ngbs_density[MAX_NUM_OF_NEIGHBOURS_STARS];
 
   /*! Number of interactions in the force SELF and PAIR */
   int num_ngb_force;
 
   /*! List of interacting particles in the force SELF and PAIR */
-  long long ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS_STARS];
+  int64_t ids_ngbs_force[MAX_NUM_OF_NEIGHBOURS_STARS];
 #endif
 
 } SWIFT_STRUCT_ALIGN;
